Add AppUser model to the project
This commit adds a new model AppUser model to the project, which is
basically an extension of the built in User model. The model contains a
OneToOneField reference to the User model. This functionality is
implemented by creating a instance of each model in one go inside the
restore_object() method of the AppUserSerializer class. All of the
previous references to the User model have been updated and tested.

-Added IsUserOrReadOnly permission
-Added AppUser model
-Renamed and modified UserSerializer to AppUserSerializer

As a side note, removed unnecessary pre_save() methods in ThreadDetail
and CommentDetail views.