Add functionality for creating dynamic celery tasks
-tasks.py which contains tasks for evaluating which tasks need to be
bumped in geo rank for each of the geo_rank values. The formula is based
getting the number of upvotes divided by a powered value of time since
posted. Only the top 1/4 of posts will get bumped to higher visibility

-updated settings.py to hold values for how often we want to run the
periodic tasks that evaluate the ranking

-added new model which holds a reference to PeriodicTask model of
djcelery. Through this model, we schedule tasks dynamically.

-added additional logic into Thread.save() to detect whether we need to
create a new periodic task based on the longitude and latitude
information, if it is a city/state/country that does not have one
associated to it.