Add token authentication to server
This commit adds a new POST hook:

/api-token-auth/

It requires both "username" and "password" attributes values in json
format to be POSTed. If successfully authenticated, it will return a
"token" value in json format with the token to be used. The token needs
to be used with every HTTP request as follows:

GET http://127.0.0.1:8000/api/example/ -H 'Authorization: Token
9944b09199c62bcf9418ad846dd0e4bbdfc6ee4b'

That is, the 'Authorization' HTTP header needs to be modified to
include:

Authorization: Token 9944b09199c62bcf9418ad846dd0e4bbdfc6ee4b

With whitespaces separating the two strings.

For more information, see:

http://www.django-rest-framework.org/api-guide/authentication/#tokenauthentication